+++ app/backend/src/main/java/com/dcg/mvc/comment/CommentController.java
@@ -15,7 +15,7 @@ public class CommentController {
 
     @PostMapping("/lectures/{lectureId}")
     public ResponseEntity<Comment> addCommentToLecture(
-            @PathVariable int lectureId,
+            @PathVariable Long lectureId,
             @RequestBody Comment comment,
             Authentication authentication) {
         try {
