+++ app/backend/src/main/java/com/dcg/mvc/history/CourseActionHistoryRepository.java
@@ -2,6 +2,6 @@ package com.dcg.mvc.history;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseActionHistoryRepository extends JpaRepository<CourseActionHistory, Integer> {
+public interface CourseActionHistoryRepository extends JpaRepository<CourseActionHistory, Long> {
 
 }
