+++ app/backend/src/main/java/com/dcg/mvc/unit/UnitRepository.java
@@ -2,5 +2,5 @@ package com.dcg.mvc.unit;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnitRepository extends JpaRepository<Unit,Integer> {
+public interface UnitRepository extends JpaRepository<Unit,Long> {
 }
