+++ app/backend/src/main/java/com/dcg/mvc/user/UserRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String username);
     @Query(value = "select count(*) from user",nativeQuery = true)
-    public int getUserCount();
+    public Long getUserCount();
 }
