+++ app/backend/src/main/java/com/dcg/mvc/user/UserService.java
@@ -55,7 +55,7 @@ public class UserService {
         return userRepository.findByEmail(user.getEmail()).get().getFullName();
     }
 
-    public int getUserCount(){
+    public Long getUserCount(){
         return userRepository.getUserCount();
     }
